WI #2735 Refactor completion for procedure parameter
