+++ TypeCobol.LanguageServer/Completion Factory/CompletionFactory.cs
@@ -12,217 +12,6 @@ namespace TypeCobol.LanguageServer
 {
     public static class CompletionFactory
     {
-        #region Procedure Completion 
-        
-        public static List<CompletionItem> GetCompletionForProcedureParameter(Position position, FileCompiler fileCompiler, CodeElement codeElement, Token userFilterToken, Token lastSignificantToken, FunctionDeclaration procedureSignatureContext)
-        {
-            var completionItems = new List<CompletionItem>();
-            var arrangedCodeElement = codeElement as CodeElementWrapper;
-            var node = CompletionFactoryHelpers.GetMatchingNode(fileCompiler, codeElement);
-            if (node == null)
-                return completionItems;
-
-            var userFilterText = userFilterToken == null ? string.Empty : userFilterToken.Text;
-
-            //Get procedure name or qualified name
-            string procedureName = CompletionFactoryHelpers.GetProcedureNameFromTokens(arrangedCodeElement?.ArrangedConsumedTokens);
-            IEnumerable<FunctionDeclaration> calledProcedures = null;
-
-            if (procedureSignatureContext == null ||
-                !(procedureSignatureContext.QualifiedName.ToString().Equals(procedureName, StringComparison.InvariantCultureIgnoreCase) 
-                  || procedureSignatureContext.Name.Equals(procedureName, StringComparison.InvariantCultureIgnoreCase)))
-            {
-                //Try to get procedure by its name
-                calledProcedures =
-                    node.SymbolTable.GetFunctions(
-                        p =>
-                            p.Name.Equals(procedureName, StringComparison.InvariantCultureIgnoreCase) ||
-                            p.VisualQualifiedName.ToString().Equals(procedureName, StringComparison.InvariantCultureIgnoreCase),
-                            SymbolTable.Scope.Intrinsic
-                        );
-            }
-            else
-            {
-                //If the procedure name is equivalent to the signature selected by signature help, we can assume the user is always on the same procedure. 
-                calledProcedures = new List<FunctionDeclaration> {procedureSignatureContext};
-            }
-
-            var alreadyGivenTokens = arrangedCodeElement?.ArrangedConsumedTokens
-                .SkipWhile(t => t != lastSignificantToken).Skip(1)
-                .TakeWhile(t => t.TokenType != TokenType.OUTPUT && t.TokenType != TokenType.IN_OUT)
-                .Except(new List<Token>() { userFilterToken })
-                .Where(t => (t.StartIndex < position.character && t.Line == position.line + 1) || t.Line < position.line + 1);
-
-            int alreadyGivenParametersCount = 0;
-            TokenType? previousTokenType = null;
-
-            //Loop that allows to ignore qualified name parameters. 
-            if (alreadyGivenTokens != null)
-                foreach (var givenToken in alreadyGivenTokens)
-                {
-                    if (givenToken.TokenType == TokenType.UserDefinedWord && (previousTokenType == null || previousTokenType.Value == TokenType.UserDefinedWord))
-                        alreadyGivenParametersCount++;
-                    previousTokenType = givenToken.TokenType;
-                }
-
-            
-            ISet<DataDefinition> potentialVariablesForCompletion = null;
-            int maxInput = -1;
-
-            int maxInOut = -1;
-            bool allProcsHaveInOutParams = true;
-            bool noProcsHaveInOutParams = true;
-            
-            bool allProcsHaveOutputParams = true;
-
-            foreach (var procedure in calledProcedures)
-            {
-                IEnumerable<ParameterDescription> procParams;
-
-                //Switch to select the correct parameters profile
-                #region Selective parameters Switch
-                switch (lastSignificantToken.TokenType)
-                {
-                    case TokenType.INPUT:
-                        {
-                            procParams = procedure.Profile.InputParameters;
-                            //Get number of input parameters left per procedure
-                            break;
-                        }
-                    case TokenType.OUTPUT:
-                        {
-                            procParams = procedure.Profile.OutputParameters;
-                            break;
-                        }
-                    case TokenType.IN_OUT:
-                        {
-                            procParams = procedure.Profile.InoutParameters;
-                            //Get number of inout parameters left per procedure
-                            break;
-                        }
-                    default:
-                        procParams = new List<ParameterDescription>();
-                        break;
-                }
-                #endregion
-
-                if (procedure.Profile.InputParameters.Count > maxInput) {
-                    maxInput = procedure.Profile.InputParameters.Count;
-                }
-                if (procedure.Profile.InoutParameters.Count > maxInOut) {
-                    maxInOut = procedure.Profile.InoutParameters.Count;
-                }
-
-                if (procedure.Profile.InoutParameters.Count == 0) {
-                    allProcsHaveInOutParams = false;
-                } else {
-                    noProcsHaveInOutParams = false;
-                }
-                if (procedure.Profile.OutputParameters.Count == 0) {
-                    allProcsHaveOutputParams = false;
-                }
-                
-                
-                //If the user already written all or more parameters than required let's check for an other proc signature
-                if (alreadyGivenParametersCount >= procParams.Count())
-                    continue;
-
-                //Else see which parameter could be filled
-                var parameterToFill = procParams.ToArray()[alreadyGivenParametersCount];
-                //Get local/global variable that could correspond to the parameter
-
-                potentialVariablesForCompletion = node.SymbolTable.GetVariablesByType(parameterToFill.DataType, potentialVariablesForCompletion, SymbolTable.Scope.Program);
-
-            }
-
-            if (potentialVariablesForCompletion == null) return completionItems;
-
-            //Convert potential variables into actual CompletionItems
-            var variables = potentialVariablesForCompletion
-                .Where(v => v.Name.StartsWith(userFilterText, StringComparison.InvariantCultureIgnoreCase)) //Filter on user text
-                .Distinct()
-                .SelectMany(d => node.SymbolTable.GetVariablesExplicitWithQualifiedName(new URI(d.Name)));
-            var items = CompletionFactoryHelpers.CreateCompletionItemsForVariableSetAndDisambiguate(variables, fileCompiler.CompilerOptions);
-            completionItems.AddRange(items);
-
-            CompletionFactoryHelpers.Case textCase = CompletionFactoryHelpers.GetTextCase(codeElement.ConsumedTokens.First(t => t.TokenType == TokenType.CALL).Text);
-            Dictionary<ParameterDescription.PassingTypes, string> paramWithCase = CompletionFactoryHelpers.GetParamsWithCase(textCase);
-            //If signature of procedure is available
-            if (procedureSignatureContext != null)
-            {
-                //Add IN-OUT or OUTPUT after INPUT ?
-                if (lastSignificantToken.TokenType == TokenType.INPUT
-                    && alreadyGivenParametersCount == (procedureSignatureContext.Profile.InputParameters.Count - 1))
-                {
-                    if (procedureSignatureContext.Profile.InoutParameters.Count != 0)
-                    {
-                        AddIn_OutSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-
-                    }
-                    else if (procedureSignatureContext.Profile.OutputParameters.Count != 0)
-                    {
-                        AddOutputSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-                    }
-                }
-
-                //Add OUTPUT after IN-OUT ?
-                else if (lastSignificantToken.TokenType == TokenType.IN_OUT
-                         && alreadyGivenParametersCount == (procedureSignatureContext.Profile.InoutParameters.Count - 1)
-                         && procedureSignatureContext.Profile.OutputParameters.Count != 0)
-                {
-                    AddOutputSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-                }
-            }
-            else
-            {
-                //Add IN-OUT or OUTPUT after INPUT ?
-                //If we reach the last INPUT parameter
-                if (lastSignificantToken.TokenType == TokenType.INPUT && alreadyGivenParametersCount == maxInput - 1)
-                {
-                    //If all procs have IN-OUT params
-                    if (allProcsHaveInOutParams)
-                    {
-                        AddIn_OutSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-                    }
-                    //If no procedures have IN-OUT params and all have OUTPUT params
-                    else if (noProcsHaveInOutParams && allProcsHaveOutputParams)
-                    {
-                        AddOutputSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-                    }
-                    //Otherwise we cannot choose between IN-OUT, OUTPUT and nothing, so we choose nothing and let the user add the good keyword manually.
-                    //#908 will change this behavior by asking for the signature context
-                }
-
-                //Add OUTPUT after IN-OUT ?
-                else if (lastSignificantToken.TokenType == TokenType.IN_OUT && alreadyGivenParametersCount == (maxInOut - 1))
-                {
-                    //If all procedures have OUTPUT parameter
-                    if (allProcsHaveOutputParams)
-                    {
-                        AddOutputSuffixToCompletionItems(lastSignificantToken, completionItems, paramWithCase);
-                    }
-                }
-            }
-
-            return completionItems;
-        }
-
-        private static void AddIn_OutSuffixToCompletionItems(Token lastSignificantToken, List<CompletionItem> completionItems, Dictionary<ParameterDescription.PassingTypes, string> paramWithCase)
-        {
-            //Use -1, because it seems LSP start counting at 1
-            var suffix = "\n" + new string(' ', lastSignificantToken.Column - 1) + paramWithCase[ParameterDescription.PassingTypes.InOut] + " ";
-            completionItems.ForEach(ci => ci.insertText += suffix);
-        }
-        private static void AddOutputSuffixToCompletionItems(Token lastSignificantToken, List<CompletionItem> completionItems, Dictionary<ParameterDescription.PassingTypes, string> paramWithCase)
-        {
-            //Use -1, because it seems LSP start counting at 1
-            var suffix = "\n" + new string(' ', lastSignificantToken.Column - 1) + paramWithCase[ParameterDescription.PassingTypes.Output] + " ";
-            completionItems.ForEach(ci => ci.insertText += suffix);
-        }
-
-
-        #endregion
-
         #region Variable Completion
         public static List<CompletionItem> GetCompletionForVariable(FileCompiler fileCompiler, CodeElement codeElement, Func<DataDefinition, bool> predicate)
         {
